cleaned sudoku_preprocessing, added sudoku_splitter
