+++ sudoku_digitalisation/features/preprocessing.py
@@ -1,17 +1,11 @@
+import datasets
 import matplotlib.pyplot as plt
 import numpy as np
 import cv2
 from PIL import Image, ImageOps
 from tqdm import tqdm
 from collections import Counter
-from datasets import load_dataset
-
-from sklearn import svm
-from sklearn.model_selection import train_test_split
-from sklearn.metrics import accuracy_score
-
-
-dataset = load_dataset("Lexski/sudoku-image-recognition", split="train")
+from datasets import load_dataset, DatasetDict, Dataset
 
 
 def load_data(path):
@@ -64,6 +58,7 @@ def crop_to_bounding_box(ds, output_size=(450, 450)):
 
     return ds
 
+
 def adaptive_histogram_equalization(ds):
     """
     Applies adaptive histogram equalization to the image.
@@ -75,26 +70,12 @@ def adaptive_histogram_equalization(ds):
     ds['image'] = Image.fromarray(equalized_image)
     return ds
 
-def hough_transform(ds):
-    """
-    Applies Hough transform to the image.
-    This is done using cv2.HoughLinesP.
-    """
-    image = np.array(ds['image'])
-    edges = cv2.Canny(image, 50, 150, apertureSize=3)
-    lines = cv2.HoughLinesP(edges, 1, np.pi / 180, threshold=100, minLineLength=500, maxLineGap=100)
-    if lines is not None:
-        for line in lines:
-            x1, y1, x2, y2 = line[0]
-            cv2.line(image, (x1, y1), (x2, y2), (0, 255, 0), 2)
-    ds['image'] = Image.fromarray(image)
-    return ds
 
-def split_image(ds):
+def split_image(image):
     """
     Splits the image into 81 cells.
     """
-    image = np.array(ds['image'])
+    image = np.array(image)
     cells = []
     cell_size = (image.shape[0] // 9, image.shape[1] // 9)
     for i in range(9):
@@ -104,39 +85,62 @@ def split_image(ds):
             y_start = j * cell_size[1]
             y_end = (j + 1) * cell_size[1]
             cell = image[x_start:x_end, y_start:y_end]
-            cells.append(cell)
-    ds['cells'] = cells
-    return ds
+            cells.append(Image.fromarray(cell))
+    return cells
+
+
+def change_label(label):
+    """Changes the label of a single cell from binary representation to the simple integer."""
+    label[0] = abs(label[0] - 1)  # changes 0 to 1 and 1 to 0, such that below if the cell is unsolved it returns 0
+    for idx, item in enumerate(label):
+        if item == 1:
+            return idx
+    return 0
+
+
+def split_labels(labels):
+    """Splits the labels into a list of 81 labels."""
+    new_labels = []
+    for i in range(9):
+        for j in range(9):
+            new_label = change_label(labels[i][j])
+            new_labels.append(new_label)
+    return new_labels
+
+
+def create_digit_ds(ds):
+    """Sets up the datasets which we use for training."""
+    new_ds = []
+    for idx, row in enumerate(ds):
+        digit_images = split_image(row['image'])
+        digit_labels = split_labels(row['cells'])
+        for j in range(len(digit_labels)):
+            new_ds.append({"digit_img": digit_images[j], "label": digit_labels[j], "index": idx})
+    return new_ds
     
 
 def preprocess_dataset(ds_dict):
     """Preprocesses the dataset."""
+    digit_ds_dict = DatasetDict()
     for split in ds_dict:
         print(split)
         ds_dict[split] = ds_dict[split].map(crop_to_bounding_box)
         ds_dict[split] = ds_dict[split].map(convert_to_grayscale)
         ds_dict[split] = ds_dict[split].map(adaptive_histogram_equalization)
-        ds_dict[split] = ds_dict[split].map(split_image)
 
-    return ds_dict
+        digit_ds = Dataset.from_list(create_digit_ds(ds_dict[split]))
+        digit_ds_dict[split] = digit_ds
+
+    return ds_dict, digit_ds_dict
 
 
 if __name__ == '__main__':
     dataset_dict = load_data("Lexski/sudoku-image-recognition")
-    ds_dict = preprocess_dataset(dataset_dict)
+    dataset_dict, digit_dataset_dict = preprocess_dataset(dataset_dict)
+    # digit_dataset_dict is what we use for training
 
-    # Show one of the cells
-    sample_cell = ds_dict['train'][0]['cells'][0]
-    show_image(sample_cell)
-
-    # To be implemented: cell images matched with labels
-
-
-
-
-    
-    
-    
-
-
-
+    # to see how it looks
+    show_image(dataset_dict["test"]["image"][0])
+    for i in range(12):
+        show_image(digit_dataset_dict["test"]["digit_img"][i])
+        print(digit_dataset_dict["test"]["label"][i])
